+++ app/sentry-example-page/metadata.ts
@@ -1,0 +1,6 @@
+import type { Metadata } from "next";
+
+export const metadata: Metadata = {
+  title: "sentry-example-page",
+  description: "Test Sentry for your Next.js app!",
+};